consolidate all models into models.py
