+++ fpan/admin.py
@@ -1,6 +1,5 @@
 from django.contrib.gis import admin
-from fpan.models.region import Region
-from fpan.models.managedarea import ManagedArea
+from fpan.models import Region, ManagedArea
 
 class ManagedAreaAdmin(admin.GeoModelAdmin):
