+++ fpan/models.py
@@ -1,10 +1,22 @@
 from django.contrib.gis.db import models
 
+
+class Region(models.Model):
+    name = models.CharField(max_length=254)
+    region_code = models.CharField(max_length=4)
+    geom = models.MultiPolygonField()
+
+    # Returns the string representation of the model.
+    def __str__(self):              # __unicode__ on Python 2
+        return self.name
+
+from django.contrib.gis.db import models
+
 class ManagedArea(models.Model):
     """this model functions similar to the Region model. it should be moved to a new file
     or something... preferably Region and MananagedArea would be just put in models.py
     to improve the clarity of import statements, etc."""
-    
+
     AGENCY_CHOICES = (
         ("FL Dept. of Environmental Protection, Div. of Recreation and Parks","FL Dept. of Environmental Protection, Div. of Recreation and Parks"),
         ("FL Dept. of Agriculture and Consumer Services, Florida Forest Service","FL Dept. of Agriculture and Consumer Services, Florida Forest Service"),
@@ -12,7 +24,7 @@ class ManagedArea(models.Model):
         ("FL Dept. of Environmental Protection, Florida Coastal Office","FL Dept. of Environmental Protection, Florida Coastal Office"),
         ("FL Dept. of Environmental Protection, Office of Water Policy","FL Dept. of Environmental Protection, Office of Water Policy"),
     )
-        
+
     CATEGORY_CHOICES = (
         ("State Park","State Park"),
         ("State Forest","State Forest"),
@@ -20,7 +32,7 @@ class ManagedArea(models.Model):
         ("Aquatic Preserve","Aquatic Preserve"),
         ("Water Management District","Water Management District"),
     )
-    
+
     WMD_DISTRICT_CHOICES = (
         ("North","North"),
         ("North Central","North Central"),
