--- fpan/models/__init__.py
@@ -1,4 +1,0 @@
-## allows simple import statements like
-## from fpan.models import Region
-from .region import Region
-from .managedarea import ManagedArea
